feature: initial plot toolbar
