+++ docs/pages/components/plot-toolbar.md
@@ -1,0 +1,22 @@
+---
+meta:
+    title: Plot Toolbar
+    description:
+layout: component
+---
+
+```html:preview
+<terra-plot-toolbar></terra-plot-toolbar>
+```
+
+## Examples
+
+### First Example
+
+TODO
+
+### Second Example
+
+TODO
+
+[component-metadata:terra-plot-toolbar]
