+++ src/components/plot-toolbar/plot-toolbar.ts
@@ -1,0 +1,12 @@
+import TerraPlotToolbar from './plot-toolbar.component.js';
+
+export * from './plot-toolbar.component.js';
+export default TerraPlotToolbar;
+
+TerraPlotToolbar.define('terra-plot-toolbar');
+
+declare global {
+    interface HTMLElementTagNameMap {
+      'terra-plot-toolbar': TerraPlotToolbar;
+    }
+}
