+++ src/components/plot-toolbar/plot-toolbar.types.ts
@@ -1,0 +1,1 @@
+export type MenuNames = 'download' | 'help' | 'information' | 'jupyter' | null
