+++ src/components/time-series/time-series.types.ts
@@ -34,6 +34,4 @@ export type TimeSeriesMetadata = {
     [key: string]: string | number
 }
 
-export type MenuNames = 'download' | 'help' | 'information' | 'jupyter' | null
-
 export type MaybeBearerToken = string | null
