+++ src/terra_ui_components/plot_toolbar/__init__.py
@@ -1,0 +1,3 @@
+from .plot_toolbar import TerraPlotToolbar
+
+__all__ = ["TerraPlotToolbar"]
